Frame improvements: pack_len and string for BackwardFrames
Add a pack_len() method to Frame to convert frames to fixed-length byte
strings.  This is useful for device drivers where the device has a
fixed-length field for frames.

Change the string output for BackwardFrame and BackwardFrameError to be
more readable and less redundant: for a frame of value "55" for example,
output BackwardFrame(55) instead of BackwardFrame(8,[55])