Response changes: raw_value property and NumericResponse
Add a 'raw_value' property to Response objects, giving access to the
underlying frame (or None for a 'No' response) without any processing.

Introduce two new Response subtypes: NumericResponse and
NumericResponseMask.  Accessing the 'value' property of these never
raises an exception, and the result is as follows:

 - '(missing)' if there was no response
 - '(framing error)'
 - integers 0 through 255 for NumericResponse
 - integers 0 through 254 and string 'MASK' for NumericResponseMask

This commit doesn't introduce any uses for these changes.