Import submodules in dali/gear/__init__.py
If some of the submodules in dali.gear are not imported, then
dali.command.from_frame() won't be able to parse their commands when
observed on the wire.  Import all of them when dali.gear is imported.

(This removes the need to import all of them explicitly in the
test_command module.)