asyncio-based drivers for USB devices that present as HID
The Tridonic DALI USB and hasseb DALI Master are USB devices that belong
to the USB "Human Interface Device" class.  They each have a HID report
descriptor that describes their packets as "manufacturer specific".

This is useful, because it means the devices can be accessed at the
"hidraw" level rather than the direct USB level, on most operating
systems.

This commit includes drivers for both devices that make use of the
hidraw interface on Linux.  There's a udev rules file in 'examples' that
will make any connected devices accessible to members of group
'dialout', and create useful symlinks under /dev/dali.

The drivers are based on the python3 asyncio library.  This should make
it simpler to write code that glues together drivers and network
services, for example using the 'gmqtt' MQTT library.

The Tridonic driver includes the ability to watch a DALI bus and create
Command and Response objects from the data observed on the wire.
examples/asyncio-buswatch.py is a simple application that makes use of
this to print DALI traffic dumps.

Extending this to work on Windows and MacOS should be possible, probably
with the introduction of a lightweight internal hidraw API.