Declare memory values as classes rather than class instances
Memory values were declared using the following pattern:

"docstring"
ValueName = MemoryValue(
    "ValueName", ...)

This introduces unnecessary repetition and the possibility for these
to be out-of-sync owing to typoes. This commit changes the way
memory values are declared to remove that repetition:

class ValueName(MemoryValue):
    "docstring"
    ...

This has some additional benefits. The documentation strings that
were previously being discarded are now accessible as ValueName.__doc__

It is easier to support memory values with unique sets of values.
There's now no need to declare a special MemoryValue subclass and
then create an instance of it; the code that handles the unique set
of values can simply be included in the memory value declaration.
For example, see oem.LightDistributionType