Remove support for multi-bank memory values
There was preliminary support in the code for memory values that
have locations spanning more than one memory bank.

I'm removing this support for the following reasons:

 * there are no memory values of this kind

 * the support was incomplete: there were several places where
   assumptions were made about memory bank based on the first or
   last location in the value, and memory values crossing banks
   would return incorrect values when using MemoryBank.read_all()

 * memory values crossing banks interacts badly with the code for
   locking and latching

Memory bank is now an attribute of memory values, rather than of
each location within them. MemoryLocation objects can no longer be
sorted, and are not sorted during read() operations; they must be
specified in the correct order in the 'locations' attribute of
memory values. (In practice, they all already were.)

I may revisit this if anybody is crazy enough to publish a
specification that requires memory values to span multiple banks!