Test MemoryValue subclasses using test data rather than default contents
Add some manually-created valid memory bank contents to the test suite,
and test that all the MemoryValue subclasses are able to retrieve the
expected values. (This is how I found the BinaryValue bug in the
previous commit.)

While working on this it became apparent that scaled values with scale
factors less than 1 didn't work very well as floats; there was noticable
loss of precision even with scale factor 0.000001 (such that tests for
equality with the expected value would fail, eg. (pow(10,-6)*pow(10,5))
== 0.1 evaluates to False rather than the expected True). I've switched
the scale factors to be Decimals instead, so the returned values are
Decimals rather than floats.

ScaledNumericValue is moved into the 'energy' module because that's the
only place it's used (and that part of the DiiA specification is the
only place it's mentioned).