Frame exception updates
Change Frame to use int.from_bytes() and int.to_bytes() internally, as
discussed in #93. This changes the exceptions raised in the case of bad
parameters.

Add a new_exceptions keyword argument to the affected methods (the Frame
initialiser, and the pack_len method) to enable us to default to the old
exceptions for now along with a DeprecationWarning. The default will be
changed after the next release.