Return FlagValue.Invalid for invalid ASCII strings
The DiiA specification "DALI Part 251" states that strings for luminaire
colour and luminaire identification should be "ascii character strings",
Null terminated if shorter than defined length. It also states that the
range of values for these memory locations is [0, 0xFF].

This is inconsistent; ASCII is a 7-bit character set and values in the
range [0x80, 0xFF] are not valid.

If values in the range [0x80, 0xFF] are encountered when reading a
string from a bus unit, StringValue will return FlagValue.Invalid
instead of a string.

It's possible that this interpretation of the standard is incorrect, and
they really meant something like "UTF-8", but until they actually say so
this is the best I can do!