Add "level" support to fakes.py
This feature branch adds the necessary message handling to fakes.py to
handle the various DALI level control messages, e.g. DAPC,
RecallMaxLevel, etc. Such a feature was found to be helpful when
testing out higher level code which uses this library, by enabling the
use of fake DALI devices which correctly respond to certain messages.

Perhaps this isn’t the ideal place for such code to live, but since
fakes.py already implemented the majority of the necessary features
this seems to be the best fit within the existing structure.

Tests have been added to cover all of the additional code.