Add initial support for Lunatone LUBA interface
Lunatone make a range of devices which use their 'LUBA' protocol to
interface a DALI bus to a RS232 serial connection. This commit adds
initial support for using such a LUBA device in python-dali.

A simple example has been added, to demonstrate basic functionality. The
driver API is very similar to the existing 'hid' driver classes.