Add 'fakes' and tests for serial driver
It is helpful when testing other parts of the library to be able to use
a 'fake' device, which doesn't actually rely on any hardware.
DriverSerialDummy implements such a fake device, which writes any DALI
commands to a log file rather than over a real bus. There is limited
support for emulating DALI control gear, to use as a testing target.

Unit tests have also been added, to test the dummy implementation -
which by extension also performs at least some basic tests on the
DriverSerialBase class.