Use gear/device address knowledge in memory objects
The only difference between memory banks in DALI 16-bit control gear and
24-bit control devices is the messages used - the banks themselves are
accessed and structured in the same way. Previously the classes for
dealing with memory assumed that a DALI 16-bit gear was being used, this
commit makes the read/write methods aware of the type of equipment and
uses the appropriate commands as required.

To read from a control gear, the API for the memory classes is the same
- i.e. an integer can be used as the address, or to be explicit a
GearShort object can be used. To read from a control device, a
DeviceShort address object must be used.