Use quiescent mode during control device scan
When scanning the bus for DALI control devices in
DeviceInstanceTypeMapper autodiscover() it is helpful to enable
"quiescent" mode first. This mode will stop control devices from sending
out any message over the DALI bus except for responses to queries, which
is helpful during this configuration process to reduce bus contention
and errors.