tridonic: recover when the device doesn't report back the response to our packet
When we issue a command via the Lunatone DALI-to-USB dongle, the first
USB packet which we get back has a copy of what we sent, like this:

 20:13:24 D: tridonic: _raw_write: 12D6000300002A01000000
 20:13:24 D: tridonic: _send_raw waiting for outstanding_transmissions=1 response=None
 20:13:24 D: tridonic: _handle_read 127300002A0100F5D6
 [...]
 20:13:24 D: tridonic: _send_raw got message: mode=12 rtype=73 frame=00002a01 interval=00f5 seq=d6
 20:13:24 D: tridonic: _send_raw waiting for outstanding_transmissions=0 response=None

When stuff works as expected, this is also accompanied by a USB packet
which tells us what the USB adapter read back from DALI in response to
our command. It looks like this:

 20:13:24 D: tridonic: _handle_read 127100000000008AD6
 20:13:24 D: tridonic: _send_raw got message: mode=12 rtype=71 frame=00000000 interval=008a seq=d6
 20:13:24 D: tridonic: _send_raw finished seq=d6

Sometimes this second USB packet, however, never arrives. When that
happens, the USB dongle appears to continue working properly, but the
code would get stuck waiting for a response which will never arrive (and
holding a lock, preventing other commands form being sent).

Solve this by treating that as an erroneous response, but continue
working. In my experience, this is sufficient and quicker than a
full reconnect.

fixes #124